+++ DuDoong-Domain/src/main/java/band/gosrock/domain/domains/cart/service/CartDomainService.java
@@ -2,16 +2,21 @@ package band.gosrock.domain.domains.cart.service;
 
 
 import band.gosrock.common.annotation.DomainService;
+import band.gosrock.domain.common.aop.redissonLock.RedissonLock;
+import band.gosrock.domain.domains.cart.adaptor.CartAdaptor;
+import band.gosrock.domain.domains.cart.domain.Cart;
 import lombok.RequiredArgsConstructor;
 
 @DomainService
 @RequiredArgsConstructor
 public class CartDomainService {
 
-    //    public Cart createCart(List<CartLineItem> cartLines, Long userId) {
-    //        //TODO : 담을려는 상품의 재고가 남아있는지 확인 ?
-    //        //TODO : 옵션이랑 연관관계 짓기... 가격 계산하기?
-    //        return
-    //
-    //    }
+    private final CartAdaptor cartAdaptor;
+
+    // 한유저당 한 카트를 소유할 수 있는 제약조건을 가짐
+    @RedissonLock(LockName = "카트생성", paramClassType = Cart.class, identifier = "userId")
+    public Long createCart(Cart cart) {
+        Cart upsert = cartAdaptor.upsert(cart);
+        return upsert.getId();
+    }
 }
